Unselect any selected items when picking a tag
When filtering by tag we don't know if selected item(s) exist in
filtered results set (and because of pagination we may still not know
this once first page of results is back). Therefore it should be handled
in the same way as search, i.e. selected items are reset upon tag
navigation.